VER: Model and runscript for orthogonal waves
